07_oauth/requirements.txt
flask
requests_oauthlib
google-cloud-datastore
gunicorn